asdf

+++ DNS_Netassay/PaperResults/combined_sim_v1.py
@@ -98,16 +98,16 @@ def parse_dns_response(ip_packet, ts):
                             knownlistDicts_stages[g][q][d][0] = knownlistDicts_stages[g][q][d][0] + 1
                             
                             serverIP = socket.inet_ntoa(rr.rdata)
-                            serverIP32 = np.uint32(int.from_bytes(socket.inet_aton(serverIP), byteorder='big'))
-                            clientIP32 = np.uint32(int.from_bytes(socket.inet_aton(clientIP), byteorder='big'))
+                            serverIP32 = np.uint64(int.from_bytes(socket.inet_aton(serverIP), byteorder='big'))
+                            clientIP32 = np.uint64(int.from_bytes(socket.inet_aton(clientIP), byteorder='big'))
 
-                            salts = [np.uint32(134140211), np.uint32(187182238), np.uint32(187238), np.uint32(1853238), np.uint32(1828), np.uint32(12238), np.uint32(72134), np.uint32(152428), np.uint32(164314534), np.uint32(223823)]
+                            salts = [np.uint64(134140211), np.uint64(187182238), np.uint64(187238), np.uint64(1853238), np.uint64(1828), np.uint64(12238), np.uint64(72134), np.uint64(152428), np.uint64(164314534), np.uint64(223823)]
                             key = clientIP + serverIP
 
                             for z in range(0, 8):
 
                                 if modulo > 0:
-                                    hashz = (crc16.crc16xmodem(np.uint32(serverIP32 + clientIP32 + salts[z])) & 0xffffffff) % modulo
+                                    hashz = (crc16.crc16xmodem(np.uint64(serverIP32 + clientIP32 + salts[z])) & 0xffffffff) % modulo
                                 else:
                                     hashz = 0
 
@@ -142,9 +142,9 @@ def parse_tcp(packet_len, ip_packet, ts):
             unlimitedKnownDict[d][2] = unlimitedKnownDict[d][2] + packet_len
 
 
-    serverIP32 = np.uint32(int.from_bytes(socket.inet_aton(source), byteorder='big'))
-    clientIP32 = np.uint32(int.from_bytes(socket.inet_aton(dest), byteorder='big'))
-    salts = [np.uint32(134140211), np.uint32(187182238), np.uint32(187238), np.uint32(1853238), np.uint32(1828), np.uint32(12238), np.uint32(72134), np.uint32(152428), np.uint32(164314534), np.uint32(223823)]
+    serverIP32 = np.uint64(int.from_bytes(socket.inet_aton(source), byteorder='big'))
+    clientIP32 = np.uint64(int.from_bytes(socket.inet_aton(dest), byteorder='big'))
+    salts = [np.uint64(134140211), np.uint64(187182238), np.uint64(187238), np.uint64(1853238), np.uint64(1828), np.uint64(12238), np.uint64(72134), np.uint64(152428), np.uint64(164314534), np.uint64(223823)]
 
     for g in [1, 2, 4, 8]:
         for q in range(0, 34, 2):
@@ -161,7 +161,7 @@ def parse_tcp(packet_len, ip_packet, ts):
                     knownlistDicts_stages[g][q][d][2] = knownlistDicts_stages[g][q][d][2] + packet_len
                     
                     if modulo > 0:
-                        hashz = (crc16.crc16xmodem(np.uint32(serverIP32 + clientIP32 + salts[z])) & 0xffffffff) % modulo
+                        hashz = (crc16.crc16xmodem(np.uint64(serverIP32 + clientIP32 + salts[z])) & 0xffffffff) % modulo
                     else:
                         hashz = 0
